+++ lib/pages/home_page.ex
@@ -5,8 +5,12 @@ defmodule Soonex.HomePage do
   use Corex
 
   import Soonex.HomePage.Hero, only: [hero: 1]
+  import Soonex.HomePage.Contents, only: [contents: 1]
+  import Soonex.HomePage.Principia, only: [principia: 1]
+  import Soonex.HomePage.Capita, only: [capita: 1]
+  import Soonex.HomePage.Kalendarium, only: [kalendarium: 1]
+  import Soonex.HomePage.Voces, only: [voces: 1]
   import Soonex.HomePage.Highlights, only: [highlights: 1]
-  import Soonex.HomePage.Scale, only: [scale: 1]
   import Soonex.HomePage.Faq, only: [faq: 1]
   import Soonex.HomePage.Waitlist, only: [waitlist: 1]
 
@@ -17,7 +21,6 @@ defmodule Soonex.HomePage do
         :countdown_ms,
         max(DateTime.diff(~U[2026-09-01 00:00:00Z], DateTime.utc_now(), :millisecond), 0)
       )
-      |> Map.put(:stats_components, length(Corex.component_ids()))
       |> Map.put(
         :posts,
         assigns
@@ -28,8 +31,12 @@ defmodule Soonex.HomePage do
     ~H"""
     <div id="home" class="w-full text-ink">
       <.hero countdown_ms={@countdown_ms} />
+      <.contents />
+      <.principia />
+      <.capita />
+      <.kalendarium />
+      <.voces />
       <.highlights posts={@posts} />
-      <.scale stats_components={@stats_components} />
       <.faq />
       <.waitlist />
     </div>